+++ constant/db.go
@@ -1,0 +1,7 @@
+package constant
+
+import "github.com/lib/pq"
+
+var (
+	DBDuplicateError = pq.ErrorCode("23505")
+)