+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to `svachan` will be documented in this file
 
-## 1.0.0 - 2020-08-14
+## 1.0.3 - 2020-09-07
+
+- Added Twitter Unknown Command Unit Test
+- Fix Set Member Email Should Update When Username Exist
+
+## 1.0.2 - 2020-09-06
+
+- Fix Server Disk Unit Test
+
+## 1.0.1 - 2020-09-06
+
+- Fix Server Disk Used Info
+
+## 1.0.0 - 2020-09-06
 
 - Initial Release